refactor: replace spectral linting with AJV
SL-2295
