+++ package.json
@@ -52,7 +52,8 @@
     "test": "jest",
     "test.prod": "yarn build && yarn lint && yarn test --coverage --maxWorkers=2",
     "test.update": "yarn test --updateSnapshot",
-    "test.watch": "yarn test --watch"
+    "test.watch": "yarn test --watch",
+    "schema.update": "yarn typescript-json-schema --id \"http://stoplight.io/schemas/rule.schema.json\" --required tsconfig.json IRule --out ./src/meta/rule.schema.json"
   },
   "dependencies": {
     "@oclif/command": "^1.0",
@@ -68,7 +69,8 @@
     "lodash": ">=4.17.5",
     "node-fetch": "2.x.x",
     "strip-ansi": "^5.2.0",
-    "text-table": "^0.2.0"
+    "text-table": "^0.2.0",
+    "typescript-json-schema": "^0.37.0"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1",
