+++ src/config/__tests__/__fixtures__/valid-ruleset.json
@@ -1,0 +1,11 @@
+{
+  "rules": {
+    "valid-rule": {
+      "summary": "should be OK",
+      "given": "$.info",
+      "then": {
+        "function": "truthy"
+      }
+    }
+  }
+}
