--- src/rulesets/spectral/index.ts
@@ -1,5 +1,0 @@
-import { RuleCollection } from '../../types';
-import { ruleSchema } from './ruleSchema';
-import { schemaToRuleCollection } from './schemaToRuleCollection';
-
-export const spectralRules = (): RuleCollection => schemaToRuleCollection(ruleSchema);
