+++ src/types/ruleset.ts
@@ -1,0 +1,5 @@
+import { RuleCollection } from './spectral';
+
+export interface IRuleset {
+  rules: RuleCollection;
+}
